Add the symmetric encryption GPG key
Read 64 bytes [512 bits] from /dev/random which
will serve as the data symmetric key for AES-XTS
encryption.

Save the key encrypted using GPGME. The encryption
method is symmetric AES-CFB, and it is protected
by a passhrase. The encrypted data key is saved
on a file named "secret_sym_key.gpg".