Add the first version of the encrypted framework
Additions:
1. Add start_enc function. This function does nothing useful.
It needs to be changed. It just reads 64 bytes from /dev/random
and create a symmetric key. This will be the key for the "real
data encryption/decryption".
Then it saves this key in file "secret_sym_key.gpg" using GPGME.

2. Before pread() and pwrite() in the synchronous case, we add
the decryption and encryption respectively.

This framework thus far, seems to work correctly.