Add support for receiving a previously created key
We can run demo_event in two ways:
1) ./demo_event -f <backing_file>
2) ./demo_event -f <backing_file> -e <encrypted_symmetric_key_file>

In the first case, demo_event will create a symmetric key that will
be used for the real data encryption and will ask for a passphrase.
In the second case, we pass the encrypted file, with the symmetric
key previously created and stored there. So in this case we just
open the <encrypted_symmetric_key_file>, decrypt it and store its
contents to `enc.key` field so that we can use the key latter for
our real encryption.

In this comimt we add support for the second case.